fix: colors now stay put even after reload
